Remove output_processor
